change variables types (double to float, long to int)

+++ echo.cpp
@@ -1,12 +1,12 @@
 #include "echo.hpp"
 #include "global.hpp"
 
-Echo::Echo() : delayTime((long)(SAMPLE_RATE * 0.5)), feedbackGain(0.7), ptr(0), status(OFF)
+Echo::Echo() : delayTime((int)(SAMPLE_RATE * 0.5)), feedbackGain(0.7), ptr(0), status(OFF)
 {
     Effect();
 }
 
-Echo::Echo(double delayTimeSeconds, double feedbackGain) : delayTime((long)(SAMPLE_RATE * delayTimeSeconds)), feedbackGain(feedbackGain), ptr(0), status(OFF)
+Echo::Echo(float delayTimeSeconds, float feedbackGain) : delayTime((int)(SAMPLE_RATE * delayTimeSeconds)), feedbackGain(feedbackGain), ptr(0), status(OFF)
 {
     Effect();
 }
@@ -16,12 +16,12 @@ Echo::~Echo()
 
 }
 
-void Echo::setFeedbackGain(double _feedbackGain)
+void Echo::setFeedbackGain(float _feedbackGain)
 {
     feedbackGain = _feedbackGain;
 }
 
-void Echo::setDelayTime(double _delayTimeSeconds)
+void Echo::setDelayTime(float _delayTimeSeconds)
 {
     delayTime = _delayTimeSeconds * SAMPLE_RATE;
 }
@@ -31,7 +31,7 @@ void Echo::setStatus(int _status)
     status = _status;
 }
 
-double Echo::process(double in)
+float Echo::process(float in)
 {
     input[ptr] = in;
     if(!status)
